move pixels to autocomplete class
